Export IpAddressVector and IpNetworkVector
