+++ inst/include/ipaddress/bitwise.h
