+++ src/Makevars
@@ -1,2 +1,2 @@
-PKG_CPPFLAGS += -I../inst/include/
+PKG_CPPFLAGS = -I../inst/include/
 PKG_CXXFLAGS = -DASIO_STANDALONE -DASIO_NO_DEPRECATED
