+++ src/Makevars.win
@@ -1,4 +1,4 @@
-PKG_CPPFLAGS += -I../inst/include/
+PKG_CPPFLAGS = -I../inst/include/
 
 # Windows target version: https://docs.microsoft.com/en-us/cpp/porting/modifying-winver-and-win32-winnt
 # set value corresponds to Windows 7
