+++ src/other_repr.h
@@ -1,36 +1,11 @@
-#ifndef __IPADDRESS_ENCODING__
-#define __IPADDRESS_ENCODING__
+#ifndef __IPADDRESS_OTHER_REPR__
+#define __IPADDRESS_OTHER_REPR__
 
-#include <algorithm>
 #include <bitset>
-#include <asio/ip/network_v4.hpp>
-#include <asio/ip/network_v6.hpp>
 #include <boost/multiprecision/cpp_int.hpp>
 
 
 namespace ipaddress {
-
-typedef std::array<int32_t, 1> r_address_v4_type;
-typedef std::array<int32_t, 4> r_address_v6_type;
-
-template<class RAddress, class CAddress>
-RAddress encode_r(const CAddress &x) {
-  RAddress x_out;
-  typename CAddress::bytes_type x_in = x.to_bytes();
-
-  std::memcpy(x_out.begin(), x_in.begin(), x_in.size());
-
-  return x_out;
-}
-
-template<class CAddress, class RAddress>
-CAddress decode_r(const RAddress &x_in) {
-  typename CAddress::bytes_type x_out;
-
-  std::memcpy(x_out.begin(), x_in.begin(), x_out.size());
-
-  return CAddress(x_out);
-}
 
 template<class Address>
 std::string encode_binary(const Address &x) {
@@ -68,7 +43,7 @@ std::string encode_integer(const Address &x, bool hex) {
     boost::multiprecision::unsigned_magnitude,
     boost::multiprecision::checked,
     void
-  >> x_int;
+    >> x_int;
 
   // import most-significant byte first (x_bytes stored in network order)
   import_bits(x_int, x_bytes.begin(), x_bytes.end(), 8, true);
@@ -91,7 +66,7 @@ Address decode_integer(const std::string& x_int_string) {
     boost::multiprecision::unsigned_magnitude,
     boost::multiprecision::checked,
     void
-  >> x_int(x_int_string);
+    >> x_int(x_int_string);
 
   // must export least-significant byte first because boost::multiprecision is
   // aggressively optimized and does not reserve space for unused bits,
